[ADD] duplicate_planning_slot
[REF] add first duplicate button

There was an issue where on certain gantt views, the duplicate button
was not showing up protperly. It was too far off to one side or was
hidden by other buttons.

This splits it so that we use the standard cell buttons for the first
cell, but keep the new slot buttons sections for the following pills.

[BUG] handle slots across multiple cells

Well dealing with a single slot element that stretches horizontally
across multiple cells, the first button group is recreated for every
cell on hover. This means that we lose reference to the actually element
being hovered. There's no correlation between the button group and the
original element.

So this caches it on hover and use it for reference

[BUG] prevent errors when trying to dup blank slots

[BUG] prevent duplicating in empty cells by accident

[BUG] handle switch into a cell with buttons open

When moving moving into a new cell, directly into a set of open button
actions, the last pill cache was still being reset. So this preventing
duplicating the first pill. Check for which div the mouse is moving into
before resetting the cache.